User API

+++ server_2.0/GYMAPP/GymApp.Infrastructure/Repositories/UserRepository.cs
@@ -1,44 +1,59 @@
-﻿using GymApp.API.DbContextModels;
-using GymApp.Core.Interfaces;
+﻿using GymApp.Core.Interfaces;
+using GymApp.Infrastructure.DbContextModels;
+using GymApp.Shared.Helpers;
 using GymApp.Shared.Models;
+using GymApp.Shared.Models.Users;
 using Microsoft.EntityFrameworkCore;
 
 namespace GymApp.Infrastructure.Repositories;
 
 public class UserRepository(ApplicationDbContext context) : IUserRepository
 {
-    public async Task<User?> GetByIdAsync(int id)
+    public async Task<User?> GetByIdAsync(Guid id)
     {
-        var user = await context.Users.FindAsync(id);
-        return user ?? null;
+        return await context.Users.FirstOrDefaultAsync(x=>x.Id == id);
     }
 
     public async Task<User?> GetByEmailAndPasswordAsync(string email, string password)
     {
-        return await context.Users
-            .FirstOrDefaultAsync(u => u.Email == email && u.Password == password); // Замените на хэшированный пароль
+        var passwordHash = HashHelper.HashPassword(password);
+        return await context.Users.Include(x=>x.EmailConfirmation)
+            .FirstOrDefaultAsync(u => u.Email == email && u.PasswordHash == passwordHash);
     }
 
-    public async Task<User> AddAsync(User? user)
+    public async Task AddAsync(User user)
     {
-        user.Id = new Guid();
-        var newUser = await context.Users.AddAsync(user);
-        await context.SaveChangesAsync(default);
-        return newUser.Entity;
+        await context.Users.AddAsync(user);
+        try
+        {
+            await context.SaveChangesAsync();
+        }
+        catch (DbUpdateException ex)
+        {
+            var dbInnerException = ex.InnerException;
+            ExeptionHelper.ThrowDbException(dbInnerException);
+        }
     }
 
     public async Task UpdateAsync(User user)
     {
         context.Users.Update(user);
+        try
+        {
+            await context.SaveChangesAsync();
+        }
+        catch (DbUpdateException ex)
+        {
+            var dbInnerException = ex.InnerException;
+            ExeptionHelper.ThrowDbException(dbInnerException);
+        }
     }
 
-    public async Task DeleteAsync(int id)
+    public async Task DeleteAsync(Guid id)
     {
         var user = await context.Users.FindAsync(id);
-        if (user != null)
-        {
-            context.Users.Remove(user);
-        }
+        if (user != null) context.Users.Remove(user);
+        await context.SaveChangesAsync();
     }
 
     public async Task SaveChangesAsync()
@@ -49,5 +64,15 @@ public class UserRepository(ApplicationDbContext context) : IUserRepository
     public Task<List<User>> GetAllAsync()
     {
         return context.Users.ToListAsync();
+    }
+
+    public async Task<User?> GetByEmailTokenAsync(string token)
+    {
+        return await context.Users.Include(x=>x.EmailConfirmation).FirstOrDefaultAsync(u => u.EmailConfirmation.EmailConfirmationToken == token);
+    }
+
+    public async Task<User?> GetByEmailAsync(string email)
+    {
+        return await context.Users.FirstOrDefaultAsync(u => u.Email == email);
     }
 }
